flask-app/requirements.txt
Flask==1.1.2
requests>=2.24.0
google-cloud-storage
python-dotenv
gunicorn
